+++ nwb_conversion_tools/utils/json_schema.py
@@ -4,6 +4,8 @@ import inspect
 import warnings
 from copy import deepcopy
 from datetime import datetime
+import yaml
+import json
 from pathlib import Path
 from typing import Optional, Union
 from typing import TypeVar
@@ -14,10 +16,52 @@ import pynwb
 FilePathType = TypeVar("FilePathType", str, Path)
 FolderPathType = TypeVar("FolderPathType", str, Path)
 OptionalFilePathType = Optional[FilePathType]
+OptionalFolderPathType = Optional[FolderPathType]
 ArrayType = Union[list, np.ndarray]
 OptionalArrayType = Optional[ArrayType]
 FloatType = Union[float, np.float]
 IntType = Union[int, np.integer]
+
+
+class NoDatesSafeLoader(yaml.SafeLoader):
+    """Custom override of yaml Loader class for datetime considerations."""
+
+    @classmethod
+    def remove_implicit_resolver(cls, tag_to_remove):
+        """
+        Remove implicit resolvers for a particular tag.
+
+        Takes care not to modify resolvers in super classes.
+        Solution taken from https://stackoverflow.com/a/37958106/11483674
+
+        We want to load datetimes as strings, not dates, because we go on to serialise as jsonwhich doesn't have the
+        advanced types of yaml, and leads to incompatibilities down the track.
+        """
+        if "yaml_implicit_resolvers" not in cls.__dict__:
+            cls.yaml_implicit_resolvers = cls.yaml_implicit_resolvers.copy()
+
+        for first_letter, mappings in cls.yaml_implicit_resolvers.items():
+            cls.yaml_implicit_resolvers[first_letter] = [
+                (tag, regexp) for tag, regexp in mappings if tag != tag_to_remove
+            ]
+
+
+NoDatesSafeLoader.remove_implicit_resolver("tag:yaml.org,2002:timestamp")
+
+
+def load_dict_from_file(file_path: FilePathType) -> dict:
+    """Safely load metadata from .yml or .json files."""
+    file_path = Path(file_path)
+    assert file_path.is_file(), f"{file_path} is not a file."
+    assert file_path.suffix in [".yml", ".json"], f"{file_path} is not a valid .yml or .json file."
+
+    if file_path.suffix == ".yml":
+        with open(file=file_path, mode="r") as stream:
+            dictionary = yaml.load(stream=stream, Loader=NoDatesSafeLoader)
+    elif file_path.suffix == ".json":
+        with open(file=file_path, mode="r") as fp:
+            dictionary = json.load(fp=fp)
+    return dictionary
 
 
 def exist_dict_in_list(d, ls):
